+++ Gameboard.java
@@ -79,7 +79,7 @@ public class Gameboard {
             Plant plant = alivePlants.get(i);
             if (!plant.isAlive()) {
                 // Clear from the board grid
-                plantBoard[plant.getRowPos()][plant.getColumnPos()] = null;
+                plantBoard[plant.getRowPos()][(int)plant.getColumnPos()] = null;
                 alivePlants.remove(i);
             }
         }
